iface: pull internal data from usid.h
Users of libsidiface_usid don't need to care about the details of how
the library communicates with the daemon, so pull that code out of
usid.h and into a new file, iface_internal.h
